requirement.txt
streamlit-shadcn-ui>=0.1.18
streamlit==1.28.1
streamlit_extras
pandas
Pillow
pyyaml